requirements.txt
amqp==2.4.2
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.2
constantly==15.1.0
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
dependencies/Model
djangorestframework==3.9.2
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
kombu==4.3.0
msgpack==0.6.1
msgpack-python==0.5.6
psycopg2==2.7.7
PyHamcrest==1.9.0
pytz==2018.9
redis==3.2.1
six==1.12.0
Twisted==18.9.0
tensorflow==1.12.0
txaio==18.8.1
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
zope.interface==4.6.0
